More refactoring for the movement thread
